requirements-agent.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
sentence-transformers>=2.5.1
langchain>=0.1.0
langchain-community>=0.0.20
langchain-huggingface>=0.0.1
faiss-cpu>=1.8.0
openai>=1.12.0
httpx>=0.27.0
pydantic>=2.6.0
